pyproject.toml
[project]
name = "llm21"
version = "0.1.0"
description = "A benchmark for evaluating how well LLMs play optimal blackjack strategy"
readme = "README.md"
license = "MIT"
authors = [{ name = "Thomas Taylor" }]
keywords = ["llm", "benchmark", "blackjack", "ai", "21"]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "jinja2>=3.1.0",
    "langchain[anthropic,google-genai,openai]>=1.1.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.32.0",
]

[project.urls]
Repository = "https://github.com/thomasgtaylor/llm21"

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
